\name{trimmedCGHcallAssays}
\alias{trimmedCGHcallAssays}
\docType{data}
\title{Sample data of class 'CGHcallAssays'}
\description{
  Encapsulates two samples from the \code{\link[CGHcall]{coriell}} data,
  but with NA values replaced by zero.
}
\usage{data(trimmedCGHcallAssays)}
\seealso{\code{\link[CGHcall]{coriell}}, \code{\link{dnacopyAssays}}}
\examples{
data(trimmedCGHcallAssays)
str(trimmedCGHcallAssays)
}
\keyword{datasets}
